Update database: New indexes and enumerate status.
The new indexes are to account for the Cron preflight hooks added in WP 5.1 so the database queries don’t suck.

A combined index is used for hook and arguments, as arguments are never queried for without specifying the hook. For hook only queries, the partial index will be used. Ideally args would be indexed using a HASH but it’s not available on enough systems.

Status can only be one of a few options.